cards refactor: - created cards container - renamed packintro to cards header
