+++ src/layout/layout.module.scss
@@ -1,6 +1,6 @@
 .wrapper {
   box-sizing: content-box;
-  max-width: 1008px;
+  max-width: 1000px;
   margin: 0 auto;
   padding: 36px 16px;
 }
