+++ src/pages/сards/cardsContainer/index.ts
